Refactor AreaTab component
